+++ frontend/src/router.ts
@@ -2,9 +2,9 @@ import Vue from 'vue';
 import Router from 'vue-router';
 
 import RouterComponent from './components/RouterComponent.vue';
-import { useUserStore } from '@/stores/user';
-import { useMainStore } from '@/stores/main';
-import { exponentialRetry } from '@/utils/job-utils';
+import {useUserStore} from '@/stores/user';
+import {useMainStore} from '@/stores/main';
+import {exponentialRetry} from '@/utils/job-utils';
 
 async function routeGuard(to, from, next) {
   const userStore = useUserStore();
@@ -316,6 +316,17 @@ export default new Router({
                           },
                         },
                       ],
+                    },
+                    {
+                      path: 'configuration',
+                      name: 'test-suite-configuration',
+                      component: () => import('./views/main/project/TestSuiteConfiguration.vue'),
+                      props: route => {
+                        return {
+                          suiteId: Number(route.params.suiteId),
+                          projectId: Number(route.params.id)
+                        };
+                      },
                     },
                   ],
                 },
